requirements.txt
pyautogui
pySerial
